add measure running time
