+++ README.md
@@ -1,12 +1,12 @@
 # 255-Group-Project-FALL21-
 Classification of news articles based on headlines and short descriptions.
 
-|Classification + tfidf|Accuracy Result|Classification + word2vec|Accuracy Result|
-|---|---|---|---|
-|1. Random Forest + tfidf                    | 0.64|6. logistic Regression + word2vec           | 0.70|
-|2. Naive Bayesian + tfidf                   | 0.67|7. Linear SVC + word2vec                    | 0.63|
-|3. logistic Regression + tfidf              | 0.62|8. K Neigbhors + word2vec                   | 0.59|
-|4. Linear SVC + tfidf                       | 0.72|||
-|5. K Neighbhors + tfidf                    | 0.58|||
+|Classification + tfidf|Accuracy Result|Running Time(Sec)|Classification + word2vec|Accuracy Result|Running Time(Sec)|
+|---|---|---|---|---|---|
+|1. Random Forest                    | 0.64|1562|6. logistic Regression           | 0.70|667|
+|2. Naive Bayesian                   | 0.67|760|7. Linear SVC                     | 0.63|3844|
+|3. logistic Regression              | 0.62|1021|8. K Neigbhors                    | 0.59|770|
+|4. Linear SVC                      | 0.72|180|9. Random Forest                   |0.59|1562|
+|5. K Neighbhors                    | 0.58|346||
