+++ classifiers/logistic_regression_word2vec.py
@@ -12,7 +12,7 @@ from sklearn.preprocessing import LabelEncoder
 import Vectorizers
 from Vectorizers.word2vec_vectorizer import Word2Vec_vectorizer
 from classifiers.classifier import Classifier
-
+import time
 
 class LogisticRegressionModel(Classifier, ABC):
     def __init__(self, X, y, text_vectorizer,
@@ -40,9 +40,12 @@ class LogisticRegressionModel(Classifier, ABC):
                   **kwargs):
         formatted_data = cls.read_file(data_file)
         print("************* Preprocessing the data started *****************")
+        # measure data pre processing time
+        start = time.time()
         text_vectorizer = vectorizer(vector_length=vector_length)
         X, y = text_vectorizer.fit_transform(data=formatted_data)
         print("************* Preprocessing the data Ended *****************")
+        print("Pre processing Complete Sec time :", time.time() - start)
         return cls(X, y, text_vectorizer, **kwargs)
 
     def train(self, save_model):
@@ -89,13 +92,18 @@ class LogisticRegressionModel(Classifier, ABC):
 
 
 if __name__ == "__main__":
+    # measure running time
+    start = time.time()
     model_lr = LogisticRegressionModel.load_data("../data/News_Category_Dataset_v2.json",
                                                  vectorizer=Word2Vec_vectorizer,
                                                  vector_length=160000, save_model=True,
                                                  model_path_location="../models",
                                                  model_name="logistic_regression_word2vec.pkl")
     model_lr.get_model_performance()
+    print("All Complete Sec time :", time.time() - start)
 
+# Pre processing Complete Sec time : 418.3803918361664
+# All Complete Sec time : 667.7726049423218
 #                 precision    recall  f1-score   support
 #
 #   BLACK VOICES       0.51      0.23      0.32       895
