+++ classifiers/random_forest_word2vec_classifier.py
@@ -1,30 +1,31 @@
 import os
 import pickle
 from abc import ABC
-from copy import deepcopy, copy
+from copy import copy
 
-import Vectorizers
-
-from classifiers.classifier import Classifier
-from sklearn.preprocessing import LabelEncoder
-from sklearn.model_selection import train_test_split
+import numpy as np
 from sklearn.metrics import classification_report
-from sklearn.neighbors import KNeighborsClassifier
+from sklearn.model_selection import train_test_split
+from sklearn.preprocessing import LabelEncoder
+import Vectorizers
+from Vectorizers.word2vec_vectorizer import Word2Vec_vectorizer
+from classifiers.classifier import Classifier
+from sklearn.ensemble import RandomForestClassifier
 import time
 
-class KNeighbhorsModel(Classifier, ABC):
+class randomforestModel(Classifier, ABC):
     def __init__(self, X, y, text_vectorizer,
                  save_model=True,
                  model_path_location="../models",
-                 model_name="Kneighbors_tfidf.pkl"):
+                 model_name="randomforest_word2vec.pkl"):
         super().__init__()
         self.X = X
         self.label_encoder = LabelEncoder()
         self.y = self.label_encoder.fit_transform(y)
         self.vectorizer = text_vectorizer
-        self.X_train, self.X_test, self.y_train, self.y_test = train_test_split(X, self.y,
+        self.X_train, self.X_test, self.y_train, self.y_test = train_test_split(X, y,
                                                                                 stratify=y,
-                                                                                test_size=0.33,
+                                                                                test_size=0.20,
                                                                                 random_state=45,
                                                                                 shuffle=True)
         self.model = self.train(save_model)
@@ -33,10 +34,9 @@ class KNeighbhorsModel(Classifier, ABC):
 
     @classmethod
     def load_data(cls, data_file,
-                  vectorizer=Vectorizers.TFIDF_vectorizer.TFIDVectorizer,
+                  vectorizer=None,
                   vector_length=150000,
                   **kwargs):
-
         formatted_data = cls.read_file(data_file)
         print("************* Preprocessing the data started *****************")
         # measure data pre processing time
@@ -49,7 +49,7 @@ class KNeighbhorsModel(Classifier, ABC):
 
     def train(self, save_model):
         print("************* model training started *****************")
-        model = KNeighborsClassifier()
+        model = RandomForestClassifier()
         model.fit(self.X_train, self.y_train)
         print("************* model training stopped *****************")
         return model
@@ -79,11 +79,11 @@ class KNeighbhorsModel(Classifier, ABC):
                 text_classifier = pickle.load(f)
             return text_classifier
         else:
-            return KNeighbhorsModel.load_data("../data/News_Category_Dataset_v2.json",
+            return randomforestModel.load_data("../data/News_Category_Dataset_v2.json",
                                                      vectorizer=Vectorizers.TFIDF_vectorizer.TFIDVectorizer,
                                                      vector_length=160000, save_model=True,
                                                      model_path_location="../models",
-                                                     model_name="Kneighbors_tfidf.pkl")
+                                                     model_name="randomforest_word2vec.pkl")
 
     def get_model_performance(self):
         y_pred = self.model.predict(self.X_test)
@@ -93,40 +93,42 @@ class KNeighbhorsModel(Classifier, ABC):
 if __name__ == "__main__":
     # measure running time
     start = time.time()
-    model_lr = KNeighbhorsModel.load_data("../data/News_Category_Dataset_v2.json",
-                                                 vectorizer=Vectorizers.TFIDF_vectorizer.TFIDVectorizer,
+    model_lr = randomforestModel.load_data("../data/News_Category_Dataset_v2.json",
+                                                 vectorizer=Word2Vec_vectorizer,
                                                  vector_length=160000, save_model=True,
                                                  model_path_location="../models",
-                                                 model_name="Kneighbors_tfidf.pkl")
+                                                 model_name="randomforest_word2vec.pkl")
     model_lr.get_model_performance()
     print("All Complete Sec time :", time.time() - start)
 
-# Pre processing Complete Sec time : 169.5145092010498
-# All Complete Sec time : 346.5285270214081
-#               precision    recall  f1-score   support
+# Pre processing Complete Sec time : 417.91098070144653
+# All Complete Sec time : 1562.4113137722015
+
+#                 precision    recall  f1-score   support
 #
-#            0       0.31      0.45      0.37      1477
-#            1       0.35      0.46      0.40      1872
-#            2       0.30      0.41      0.34      1603
-#            3       0.43      0.54      0.48      1107
-#            4       0.53      0.61      0.57      1130
-#            5       0.52      0.73      0.61      5162
-#            6       0.57      0.74      0.64      2050
-#            7       0.25      0.28      0.26      2000
-#            8       0.69      0.60      0.64      1381
-#            9       0.42      0.22      0.29      1115
-#           10       0.53      0.30      0.38      2863
-#           11       0.31      0.20      0.24      1272
-#           12       0.71      0.82      0.76     10668
-#           13       0.70      0.48      0.57      2049
-#           14       0.68      0.68      0.68      1569
-#           15       0.78      0.64      0.70      3182
-#           16       0.71      0.59      0.64      3200
-#           17       0.79      0.56      0.66      1205
-#           18       0.64      0.49      0.55      5882
-#           19       0.43      0.18      0.26      1099
-#           20       0.70      0.49      0.58      1209
+#   BLACK VOICES       0.64      0.12      0.21       895
+#       BUSINESS       0.54      0.28      0.37      1135
+#         COMEDY       0.61      0.16      0.26       972
+#          CRIME       0.57      0.54      0.56       670
+#        DIVORCE       0.71      0.47      0.56       685
+#  ENTERTAINMENT       0.49      0.75      0.59      3129
+#   FOOD & DRINK       0.66      0.68      0.67      1242
+# HEALTHY LIVING       0.56      0.06      0.11      1212
+#  HOME & LIVING       0.71      0.44      0.55       837
+#         IMPACT       0.48      0.08      0.14       676
+#      PARENTING       0.44      0.59      0.51      1735
+#        PARENTS       0.53      0.09      0.16       771
+#       POLITICS       0.67      0.89      0.77      6466
+#   QUEER VOICES       0.69      0.37      0.48      1242
+#         SPORTS       0.70      0.33      0.45       951
+# STYLE & BEAUTY       0.74      0.73      0.74      1928
+#         TRAVEL       0.59      0.73      0.65      1939
+#       WEDDINGS       0.82      0.53      0.64       730
+#       WELLNESS       0.50      0.80      0.62      3565
+#          WOMEN       0.48      0.16      0.24       666
+#      WORLDPOST       0.72      0.34      0.47       733
 #
-#     accuracy                           0.58     53095
-#    macro avg       0.54      0.50      0.51     53095
-# weighted avg       0.59      0.58      0.57     53095
+#       accuracy                           0.59     32179
+#      macro avg       0.61      0.44      0.46     32179
+#   weighted avg       0.60      0.59      0.55     32179
+#
